feat: apply button

+++ src/main/java/com/example/demo/model/Post.java
@@ -6,7 +6,7 @@ import java.math.BigInteger;
 
 @Data
 public class Post {
-    public BigInteger proposal_id;
+    public long proposal_id;
     public String user_id;
     public String title;
     public String content;
